Further backports of db_int*/phpdoc*

+++ core/tokens_api.php
@@ -44,15 +44,15 @@ $g_tokens_purged = false;
  * @return boolean True if token exists
  */
 function token_exists( $p_token_id ) {
-	$c_token_id = db_prepare_int( $p_token_id );
 	$t_tokens_table = db_get_table( 'tokens' );
+	$t_query = "SELECT id FROM $t_tokens_table WHERE id=" . db_param();
+	$t_result = db_query_bound( $t_query, array( $p_token_id ), 1 );
 
-	$t_query = "SELECT id
-		          	FROM $t_tokens_table
-		          	WHERE id=" . db_param();
-	$t_result = db_query_bound( $t_query, array( $c_token_id ), 1 );
-
-	return( 1 == db_num_rows( $t_result ) );
+	$t_row = db_fetch_array( $t_result );
+	if ( $t_row ) {
+		return true;
+	}
+	return false;
 }
 
 /**
@@ -77,8 +77,8 @@ function token_ensure_exists( $p_token_id ) {
 function token_get( $p_type, $p_user_id = null ) {
 	token_purge_expired_once();
 
-	$c_type = db_prepare_int( $p_type );
-	$c_user_id = db_prepare_int( $p_user_id == null ? auth_get_current_user_id() : $p_user_id );
+	$c_type = (int)$p_type;
+	$c_user_id = (int)( $p_user_id == null ? auth_get_current_user_id() : $p_user_id );
 
 	$t_tokens_table = db_get_table( 'tokens' );
 
@@ -86,8 +86,9 @@ function token_get( $p_type, $p_user_id = null ) {
 					WHERE type=" . db_param() . " AND owner=" . db_param();
 	$t_result = db_query_bound( $t_query, array( $c_type, $c_user_id ) );
 
-	if( db_num_rows( $t_result ) > 0 ) {
-		return db_fetch_array( $t_result );
+	$t_row = db_fetch_array( $t_result );
+	if ( $t_row ) {
+		return $t_row;
 	}
 
 	return null;
@@ -135,14 +136,10 @@ function token_set( $p_type, $p_value, $p_expiry = TOKEN_EXPIRY, $p_user_id = nu
 function token_touch( $p_token_id, $p_expiry = TOKEN_EXPIRY ) {
 	token_ensure_exists( $p_token_id );
 
-	$c_token_id = db_prepare_int( $p_token_id );
 	$c_token_expiry = time() + $p_expiry;
 	$t_tokens_table = db_get_table( 'tokens' );
-
-	$t_query = "UPDATE $t_tokens_table
-					SET expiry=" . db_param() . "
-					WHERE id=" . db_param();
-	db_query_bound( $t_query, array( $c_token_expiry, $c_token_id ) );
+	$t_query = "UPDATE $t_tokens_table SET expiry=" . db_param() . " WHERE id=" . db_param();
+	db_query_bound( $t_query, array( $c_token_expiry, $p_token_id ) );
 }
 
 /**
@@ -151,14 +148,11 @@ function token_touch( $p_token_id, $p_expiry = TOKEN_EXPIRY ) {
  * @param integer $p_user_id User ID or null for current logged in user.
  */
 function token_delete( $p_type, $p_user_id = null ) {
-	$c_type = db_prepare_int( $p_type );
 	$c_user_id = db_prepare_int( $p_user_id == null ? auth_get_current_user_id() : $p_user_id );
 
 	$t_tokens_table = db_get_table( 'tokens' );
-
-	$t_query = "DELETE FROM $t_tokens_table
-					WHERE type=" . db_param() . " AND owner=" . db_param();
-	db_query_bound( $t_query, array( $c_type, $c_user_id ) );
+	$t_query = "DELETE FROM $t_tokens_table WHERE type=" . db_param() . " AND owner=" . db_param();
+	db_query_bound( $t_query, array( $p_type, $c_user_id ) );
 }
 
 /**
@@ -169,17 +163,14 @@ function token_delete_by_owner( $p_user_id = null ) {
 	if( $p_user_id == null ) {
 		$c_user_id = auth_get_current_user_id();
 	} else {
-		$c_user_id = db_prepare_int( $p_user_id );
+		$c_user_id = (int)$p_user_id;
 	}
 
-	$t_tokens_table = db_get_table( 'tokens' );
 
 	# Remove
-	$t_query = "DELETE FROM $t_tokens_table
-		          	WHERE owner=" . db_param();
+	$t_tokens_table = db_get_table( 'tokens' );
+	$t_query = "DELETE FROM $t_tokens_table WHERE owner=" . db_param();
 	db_query_bound( $t_query, array( $c_user_id ) );
-
-	return true;
 }
 
 /**
@@ -191,10 +182,10 @@ function token_delete_by_owner( $p_user_id = null ) {
  * @return integer Token ID
  */
 function token_create( $p_type, $p_value, $p_expiry = TOKEN_EXPIRY, $p_user_id = null ) {
-	$c_type = db_prepare_int( $p_type );
+	$c_type = (int)$p_type;
 	$c_timestamp = db_now();
 	$c_expiry = time() + $p_expiry;
-	$c_user_id = db_prepare_int( $p_user_id == null ? auth_get_current_user_id() : $p_user_id );
+	$c_user_id = (int)( $p_user_id == null ? auth_get_current_user_id() : $p_user_id );
 
 	$t_tokens_table = db_get_table( 'tokens' );
 
@@ -214,11 +205,10 @@ function token_create( $p_type, $p_value, $p_expiry = TOKEN_EXPIRY, $p_user_id =
  */
 function token_update( $p_token_id, $p_value, $p_expiry = TOKEN_EXPIRY ) {
 	token_ensure_exists( $p_token_id );
-	$c_token_id = db_prepare_int( $p_token_id );
+	$c_token_id = (int)$p_token_id;
 	$c_expiry = time() + $p_expiry;
 
 	$t_tokens_table = db_get_table( 'tokens' );
-
 	$t_query = "UPDATE $t_tokens_table
 					SET value=" . db_param() . ", expiry=" . db_param() . "
 					WHERE id=" . db_param();
@@ -233,14 +223,11 @@ function token_update( $p_token_id, $p_value, $p_expiry = TOKEN_EXPIRY ) {
  * @return bool always true.
  */
 function token_delete_by_type( $p_token_type ) {
-	$c_token_type = db_prepare_int( $p_token_type );
-
-	$t_tokens_table = db_get_table( 'tokens' );
 
 	# Remove
-	$t_query = "DELETE FROM $t_tokens_table
-		          	WHERE type=" . db_param();
-	db_query_bound( $t_query, array( $c_token_type ) );
+	$t_tokens_table = db_get_table( 'tokens' );
+	$t_query = "DELETE FROM $t_tokens_table WHERE type=" . db_param();
+	db_query_bound( $t_query, array( $p_token_type ) );
 
 	return true;
 }
